Mark as paid extraction

+++ src/main/java/com/ttps/laboratorio/dto/request/UnpaidStudiesDTO.java
@@ -2,11 +2,13 @@ package com.ttps.laboratorio.dto.request;
 
 import java.util.ArrayList;
 import java.util.List;
+
+import org.springframework.validation.annotation.Validated;
+
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
-import org.springframework.validation.annotation.Validated;
 
 @Validated
 @Getter
@@ -15,6 +17,6 @@ import org.springframework.validation.annotation.Validated;
 @NoArgsConstructor
 public class UnpaidStudiesDTO {
 
-	private List<Integer> unpaidStudies = new ArrayList<>();
+	private List<Long> unpaidStudies = new ArrayList<>();
 
 }
